backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
email-validator==2.1.0.post1
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.9
chromadb>=0.4.22
sentence-transformers==2.3.1
sqlalchemy==2.0.27
groq==0.4.2
